pyproject.toml
[project]
name = "redflag-analyst"
version = "0.2.0"
description = "Deterministic red-teaming and compliance-gating engine for institutional finance"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.1,<3",
    "python-dotenv>=1.0,<2",
    "pdfplumber>=0.10,<1",
    "python-docx>=1.1,<2",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.30,<2",
    "plotly>=5.18,<7",
]
test = [
    "pytest>=8.0,<10",
    "pytest-cov>=5.0,<8",
    "fpdf2>=2.7,<3",
]

[project.scripts]
redflag = "run_redflag:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Long lines are string literals (compliance messages); readability > line length

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "app_redteam.py"]

[tool.coverage.report]
show_missing = true
fail_under = 80
